Side count changing
put in the code to change the number of sides the shape has. can be seen
in the inspector by changing the value of sides changing the bool to
true. just needs to be hooked to a slider (should only call when the
slider has changed). a coded min of 3 (otherwise, what shape) and max of
50 (performance and unneccesary detail. working alright up to 100, but
already quite circular at 20)